Port over work-in-progres castles branch.
This is a branch I started on master before the v4 split.
It is definitely not ready for prime-time! I stand by the
file splits but the whole "castles" idea may not be any
good... we'll have to see.

Even though the code here is in a transitional/debugging
state I'm committing just to avoid possibly losing data.